+++ hazelcast-sql-core/src/main/java/com/hazelcast/sql/impl/calcite/validate/operators/HazelcastCaseOperator.java
@@ -16,47 +16,63 @@
 
 package com.hazelcast.sql.impl.calcite.validate.operators;
 
-import com.hazelcast.sql.impl.calcite.validate.HazelcastCallBinding;
-import com.hazelcast.sql.impl.calcite.validate.operators.common.AbstractHazelcastCaseOperator;
+import com.hazelcast.sql.impl.QueryException;
+import com.hazelcast.sql.impl.SqlErrorCode;
+import com.hazelcast.sql.impl.calcite.validate.types.HazelcastTypeUtils;
+import com.hazelcast.sql.impl.type.QueryDataType;
+import com.hazelcast.sql.impl.type.QueryDataTypeFamily;
+import org.apache.calcite.rel.type.RelDataType;
+import org.apache.calcite.sql.SqlCall;
+import org.apache.calcite.sql.SqlKind;
+import org.apache.calcite.sql.SqlOperator;
+import org.apache.calcite.sql.SqlOperatorBinding;
+import org.apache.calcite.sql.SqlSyntax;
+import org.apache.calcite.sql.SqlWriter;
+import org.apache.calcite.sql.fun.SqlCaseOperator;
+import org.apache.calcite.sql.type.InferTypes;
 
-public final class HazelcastCaseOperator extends AbstractHazelcastCaseOperator {
+import java.util.Arrays;
+
+public final class HazelcastCaseOperator extends SqlOperator {
 
     public static final HazelcastCaseOperator INSTANCE = new HazelcastCaseOperator();
 
     private HazelcastCaseOperator() {
-        super();
+        super(SqlCaseOperator.INSTANCE.getName(), SqlKind.CASE, SqlCaseOperator.INSTANCE.getLeftPrec(), true, null,
+                InferTypes.RETURN_TYPE, null);
     }
 
     @Override
-    protected boolean checkOperandTypes(HazelcastCallBinding callBinding, boolean throwOnFailure) {
-//        return SqlCaseOperator.INSTANCE.checkOperandTypes(binding, throwOnFailure);
-//        if (binding instanceof SqlCallBinding) {
-//            SqlCallBinding sqlBinding = (SqlCallBinding) binding;
-//            SqlCase call = (SqlCase) sqlBinding.getCall();
-//            HazelcastSqlValidator validator = (HazelcastSqlValidator) sqlBinding.getValidator();
-//
-//            validator.getTypeCoercion().caseWhenCoercion(sqlBinding);
-//            RelDataType caseType = validator.getKnownNodeType(call);
-//            if (caseType == null) {
-//                throw sqlBinding.newValidationError(RESOURCE.dynamicParamIllegal());
-//            }
-//
-//            for (SqlNode thenOperand : call.getThenOperands()) {
-//                RelDataType thenOperandType = validator.deriveType(sqlBinding.getScope(), thenOperand);
-//                if (!HazelcastTypeSystem.canCast(thenOperandType, caseType)) {
-//                    throw sqlBinding.newValidationError(RESOURCE.illegalMixingOfTypes());
-//                }
-//            }
-//            SqlNode elseOperand = call.getElseOperand();
-//            RelDataType elseOperandType = validator.deriveType(sqlBinding.getScope(), elseOperand);
-//            if (!HazelcastTypeSystem.canCast(elseOperandType, caseType)) {
-//                throw sqlBinding.newValidationError(RESOURCE.illegalMixingOfTypes());
-//            }
-//
-//            return caseType;
-//        } else {
-//            return SqlCaseOperator.INSTANCE.inferReturnType(binding);
-//        }
-        return false;
+    public RelDataType inferReturnType(SqlOperatorBinding binding) {
+        int size = binding.getOperandCount();
+        RelDataType caseReturnType = binding.getOperandType(1);
+        QueryDataType firstThenBranchType = HazelcastTypeUtils.toHazelcastType(caseReturnType.getSqlTypeName());
+        QueryDataTypeFamily[] allReturnTypes = new QueryDataTypeFamily[size / 2 + 1];
+        int j = 0;
+        allReturnTypes[j] = firstThenBranchType.getTypeFamily();
+        j++;
+        boolean failure = false;
+        for (int i = 1 + 2; i < size; i += 2) {
+            QueryDataType operandType = HazelcastTypeUtils.toHazelcastType(binding.getOperandType(i).getSqlTypeName());
+            failure |= !firstThenBranchType.equals(operandType);
+            allReturnTypes[j++] = operandType.getTypeFamily();
+        }
+        QueryDataType elseType = HazelcastTypeUtils.toHazelcastType(binding.getOperandType(size - 1).getSqlTypeName());
+        failure |= !firstThenBranchType.equals(elseType);
+        allReturnTypes[j] = elseType.getTypeFamily();
+        if (failure) {
+            throw QueryException.error(SqlErrorCode.GENERIC, "Cannot infer return type of case operator among " + Arrays.toString(allReturnTypes));
+        }
+        return caseReturnType;
+    }
+
+    @Override
+    public SqlSyntax getSyntax() {
+        return SqlCaseOperator.INSTANCE.getSyntax();
+    }
+
+    @Override
+    public void unparse(SqlWriter writer, SqlCall call, int leftPrec, int rightPrec) {
+        writer.print("CASE");
     }
 }
